fix(enhance): image_histogram2d bin centers built at image.dtype
Same bug class as the already-merged bbox_to_mask fix: centers =
torch.arange(n_bins, dtype=image.dtype) built the bin-center position
grid directly at the caller's dtype. float16 can only exactly
represent integers up to 2048 (confirmed: torch.arange(4096,
dtype=torch.float16) has only 3073 distinct values, not 4096) --
with n_bins large enough, DIFFERENT bin indices silently collapse onto
the SAME center value, giving those bins numerically identical
(redundant) kernel-density output.

Fixed by building the arange at fixed float32 and deliberately NOT
casting back to image.dtype -- centers is an internal working buffer,
not a returned value, so there is no output-dtype contract to
preserve; the subsequent subtraction against image promotes normally
through ordinary type promotion. (My first attempt DID cast back to
image.dtype, matching the CoarseMatching fix's pattern for a
*returned* value -- caught before committing that it re-rounds every
center straight back down to float16 and reproduces the exact same
collapse, since this variable's role is different: internal buffer,
not output.)

Verified with a test that isolates the mechanism cleanly: compares the
function's own auto-constructed centers against an EXPLICITLY-supplied
float64 centers argument on the *same* float16 image tensor (no
cross-dtype confound from float16 also independently quantizing the
image's own pixel values). Max abs diff vs. that reference: ~5.99
unfixed, ~9e-7 (float32 rounding noise) fixed. Full enhance/ test
suite (359 tests) passes, ruff clean.

+++ kornia/enhance/histogram.py
@@ -233,7 +233,19 @@ def image_histogram2d(
         bandwidth = (max - min) / n_bins
 
     if centers is None:
-        centers = min + bandwidth * (torch.arange(n_bins, device=image.device, dtype=image.dtype) + 0.5)
+        # Build the bin-index arange at a fixed float32 regardless of image.dtype -- a
+        # low-precision image.dtype (float16/bfloat16) can't exactly represent integers past
+        # 2048 (float16) / has limited integer range (bfloat16's 8-bit mantissa), so building
+        # the arange directly at image.dtype silently collapses distinct bin indices together
+        # once n_bins is large enough: torch.arange(4096, dtype=torch.float16) has only 3073
+        # distinct values, not 4096 (confirmed directly), which collapsed centers[i] == centers[j]
+        # for i != j and gave those bins numerically identical (redundant) histogram output.
+        # Deliberately NOT cast back to image.dtype: centers is an internal working buffer, not
+        # a returned value, so there is no output-dtype contract to preserve here -- doing so
+        # would re-round every value straight back down to float16 and reproduce the exact same
+        # collapse this fix exists to prevent. `image.unsqueeze(0) - centers` below promotes
+        # normally through ordinary PyTorch type promotion instead.
+        centers = min + bandwidth * (torch.arange(n_bins, device=image.device, dtype=torch.float32) + 0.5)
     centers = centers.reshape(-1, 1, 1, 1, 1)
 
     u = torch.abs(image.unsqueeze(0) - centers) / bandwidth
